Add links to required software in lenovom73tiny-proxmox.md
The commit adds updated links to the required software (Rufus and Tera Term) in the file lenovom73tiny-proxmox.md. These links will help users easily access and download the necessary software. This change improves the user experience and avoids confusion when trying to install Proxmox on a Lenovo M73 Tiny. The changes do not impact the functionality or behavior of the code. No related issue references exist for this commit.